revize

+++ routers/todo.py
@@ -8,6 +8,7 @@ from models import Base, Todo
 from database import engine, SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
+from routers.auth import get_current_user
 
 
 router = APIRouter(
@@ -19,7 +20,7 @@ router = APIRouter(
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str = Field(min_length=3, max_length=100)
-    #priority: int = Field(gt=0, lt=6)
+    priority: int = Field(gt=0, lt=6)
     complete: bool
 
 def get_db():
@@ -30,27 +31,37 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
-@router.get("/read_all")
-async def read_all(db:db_dependency):
-    return db.query(Todo).all()
+@router.get("/")
+async def read_all(user: user_dependency, db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    return db.query(Todo).filter(Todo.owner_id == user.get('id')).all()
 
 @router.get("/get_by_id/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_by_id(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+async def read_by_id(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is not None:
         return todo
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
 
-@router.post("/create_todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency, todo_request: TodoRequest):
-    todo = Todo(**todo_request.dict())
+@router.post("/todo", status_code=status.HTTP_201_CREATED)
+async def create_todo(user: user_dependency, db:db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = Todo(**todo_request.dict(), owner_id=user.get('id'))
     db.add(todo)
     db.commit()
 
-@router.put("/update_todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency ,todo_request:TodoRequest ,todo_id:int =Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(user: user_dependency, db:db_dependency ,todo_request:TodoRequest ,todo_id:int =Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
 
@@ -61,11 +72,14 @@ async def update_todo(db:db_dependency ,todo_request:TodoRequest ,todo_id:int =P
     db.add(todo)
     db.commit()
 
-@router.delete("/delete_todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency, todo_id:int =Path(gt=0)):
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency, db:db_dependency, todo_id:int =Path(gt=0)):
     todo = db.query(Todo).filter(Todo.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+
     if todo is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
 
-    db.query(Todo).filter(Todo.id == todo_id).delete()
+    db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).delete()
     db.commit()
